// lib/import/types.ts
export interface FECColumnMapping {
  JournalCode: string
  JournalLib: string
  EcritureNum: string
  EcritureDate: string
  CompteNum: string
  CompteLib: string
  CompAuxNum?: string
  CompAuxLib?: string
  PieceRef?: string
  PieceDate?: string
  EcritureLib?: string
  Debit: string
  Credit: string
  EcritureLet?: string
  DateLet?: string
  ValidDate?: string
  Montantdevise?: string
  Idevise?: string
}